% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_fgrid-function.R
\name{aov_fgrid}
\alias{aov_fgrid}
\alias{aov_fgrid_generate}
\alias{aov_fgrid.default}
\alias{aov_fgrid}
\title{Create AOV formulas from a filter.grid}
\usage{
aov_fgrid(.data, .dep_var, .grp_var, ...)

aov_fgrid_generate(.data, .dep_var, .grp_var, ...)

aov_fgrid.default(.data, ..., .dep_var, .grp_var)

aov_fgrid(.data, .dep_var, .grp_var, ...)
}
\arguments{
\item{.data}{a dataframe or a named list of data frames}

\item{.dep_var}{a vector naming the depend. var.}

\item{.grp_var}{an optional grouping factor}

\item{...}{further arguments passed to or from other methods}
}
\value{
a data.frame with a  class attribute for passing to additional
}
\description{
Creates a dataframe of formulas for aovs from a filter grid

Create ANOVA formulas from a filter.grid data frame
}
\seealso{
Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{create_fgrid}}, \code{\link{expand_fgrid}},
  \code{\link{factor_summary}},
  \code{\link{summary_form_grid}}
}
