% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_forms-function.R
\name{aov_forms}
\alias{aov_forms}
\alias{aov_forms.default}
\alias{aov_forms.data.frame}
\alias{aov_formply}
\title{Create AOV formulas from a filter.grid}
\usage{
aov_forms(data, ...)

\method{aov_forms}{default}(data, dep_var, ind_var_btwn, grp_factor)

\method{aov_forms}{data.frame}(data, dep_var, ind_var_btwn, grp_factor)

aov_formply(data, dep_var, grp_factor, ...)
}
\arguments{
\item{data}{a \code{filter.grid} dataframe}

\item{...}{further arguments passed to or from other methods}

\item{dep_var}{a vector naming the depend. var.}

\item{ind_var_btwn}{an optional vector naming the between-subjects variable}
}
\value{
a data.frame with a \code{filter.grid} class attribute for passing to additional \link{factorial-design functions}
}
\description{
Creates a dataframe of formulas for aovs from a filter grid

A vectorized version of \code{aov_forms}

Create ANOVA formulas from a filter.grid data frame

Alternate version of \code{aov_forms}
}
\seealso{
Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_filter}},
  \code{\link{factor_grid.list}},
  \code{\link{factor_grid}}, \code{\link{factor_summary}}

Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_filter}},
  \code{\link{factor_grid.list}},
  \code{\link{factor_grid}}, \code{\link{factor_summary}}

Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_filter}},
  \code{\link{factor_grid.list}},
  \code{\link{factor_grid}}, \code{\link{factor_summary}}

Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_filter}},
  \code{\link{factor_grid.list}},
  \code{\link{factor_grid}}, \code{\link{factor_summary}}
}
