% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_fgrid_between-function.R
\name{aov_formulate_bw}
\alias{aov_formulate_bw}
\title{Helper for AOV BW formulas from a filter.grid}
\usage{
aov_formulate_bw(.data, .dep_var, .grp_var, ...)
}
\arguments{
\item{.data}{a dataframe or a named list of data frames}

\item{.dep_var}{a vector naming the depend. var.}

\item{.grp_var}{an optional grouping factor}

\item{...}{further arguments passed to or from other methods}

\item{.btw_var}{an bwteen group factor}
}
\value{
a data.frame with a  class attribute for passing to additional
}
\description{
Creates a dataframe of formulas for aovs from a filter grid
}
\seealso{
Other factorial-design functions: \code{\link{aov_fgrid}},
  \code{\link{combine_factor_levels}},
  \code{\link{create_fgrid}}, \code{\link{expand_fgrid}},
  \code{\link{factor_summary}},
  \code{\link{summary_form_grid}}
}
