% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_fstats_reshape-function.R
\name{aov_fstats_to_long}
\alias{aov_fstats_to_long}
\alias{aov_fstats_to_wide}
\title{Creates a long version of ftable}
\usage{
aov_fstats_to_long(.data, ...)

aov_fstats_to_wide(.data, .gather = TRUE, .col_ord = c("PCT", "PCTND",
  "ASN", "ASNND", "UNIN", "UNINND", "MISCEL"), ...)
}
\arguments{
\item{.data}{data to be analysed}

\item{...}{further arguments passed to or from other methods(ignored)}

\item{.gather}{gather ouput}

\item{.col_ord}{a vector giving the order of dep_var columns}
}
\value{
a data.frame with the results of an fstatstable
}
\description{
Creates a long version of ftable
}
