% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_fstats_table-function.R
\name{aov_fstats_combine}
\alias{aov_fstats_combine}
\alias{aov_fstats_table}
\title{Create AOV fstats_table}
\usage{
aov_fstats_combine(.data, stat_ls, interval, ...)

aov_fstats_table(.data, latex = TRUE, ...)
}
\arguments{
\item{.data}{data to be analysed}

\item{stat_ls}{list of stat designations for printout}

\item{interval}{should pvalues be returned as intervals?}

\item{...}{further arguments passed to or from other methods (ignored)}

\item{latex}{shoudl the output be latex?}
}
\value{
a data.frame with the results of an anova
}
\description{
Creates a table of aov summary
}
