% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_fstats_to_long-function.R
\name{aov_fstats_to_long}
\alias{aov_fstats_to_long}
\title{Creates a long version of ftable}
\usage{
aov_fstats_to_long(.data, ...)
}
\arguments{
\item{.data}{data to be analysed}

\item{...}{further arguments passed to or from other methods(ignored)}
}
\value{
a data.frame with the results of an fstatstable
}
\description{
Creates a long version of ftable
}
