% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_custom.R, R/aov_run-function.R
\name{aov_custom}
\alias{aov_custom}
\alias{aov_run}
\title{Create AOV formulas from a filter.grid}
\usage{
aov_custom(.data, .spec, ..., .custom_dv_col, .custom_aggr_funs)

aov_run(.data, .spec, ...)
}
\arguments{
\item{.data}{data to be analysed}

\item{.spec}{specs for anovas}

\item{...}{further arguments passed to or from other methods}

\item{.custom_dv_col}{names of columns of DVs (not already specified in datsel_form to use), in a list or vector}

\item{.custom_aggr_funs}{names of funs to use to aggregrate data in a vector}

\item{.data}{data to be analysed}

\item{.spec}{specs for anovas}

\item{...}{further arguments passed to or from other methods}
}
\value{
a data.frame with the results of an anova

a data.frame with the results of an anova
}
\description{
Creates a dataframe of formulas for aovs from a filter grid

Creates a dataframe of formulas for aovs from a filter grid
}
