% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_resids-function.R
\name{append_resids}
\alias{append_resids}
\title{Append residuals to a dataframe.}
\usage{
append_resids(dataframe, id.group, lm.form)
}
\arguments{
\item{dataframe}{A dataframe with at least two numeric or integer columns suitable for fitting with \code{\link[stats]{lm}}.}

\item{id.group}{A list of one or more RHS formulas (see \code{\link[lazyeval]{f_rhs}}) naming the dataframe column(s) that will be used to create a unique row identifier for binding the column of residual values to the original dataframe: list(~col1,~col2); if NULL (default), assumes each row is unique. )}

\item{lm.form}{= A list of exactly two RHS formulas (see \code{\link[lazyeval]{f_rhs}}) naming the dataframe columns over which residulas will be computed, or else a formula to be passed to  \code{\link[stats]{lm}}.}
}
\value{
A copy of the original dataframe with residual columns appended. When the \code{lm.form} arguments is a list of exactly two RHS formulas, column names are \code{col1.res.col2} & \code{col2.res.col2}; otherwise, the coloumn name is a reasonably interpretable but abbreviated version of the column names in the formula passed to \code{\link[stats]{lm}}, where the LHS and RHS are separate by \code{.res.}
}
\description{
\code{append_resids} returns a copy of dataframe with residual values from \code{\link[stats]{lm}} appended.
}
