% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc-function.R
\name{cc}
\alias{cc}
\title{Split a string at space (an alternative to \code{c()})}
\usage{
cc(string_to_split)
}
\arguments{
\item{string_to_split}{A single quoted string.}
}
\value{
A simplified character matrix
}
\description{
\code{cc} is a wrapper around \code{\link[stringr]{str_split}} that takes a single quoted string whose elements are to be split at whitespace.
}
\examples{

# compare: c("A","B","C") vs. cc("A B C")
c1 <-
  c("A","B","C")
c2 <-
   cc("A B")

c1 == c2

}

