% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_index-function.R
\name{col_index}
\alias{col_index}
\title{The names of a dataframe column in a numbered matrix.}
\usage{
col_index(X)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A numbered matrix of columnames.
}
\description{
For convenience.
}
