% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fgrid-function.R
\name{create_fgrid}
\alias{create_fgrid}
\alias{create_fgrid.default}
\alias{create_fgrid.list}
\alias{create_fgrid.data.frame}
\title{Create a data frame of variable combinations for a fully-crossed factorial design}
\usage{
create_fgrid(.data, ...)

\method{create_fgrid}{default}(.data, ...)

\method{create_fgrid}{list}(.data, ...)

\method{create_fgrid}{data.frame}(data, ..., ind_vars)
}
\arguments{
\item{.data}{data that will be converted to a factorgrid dataframe}

\item{...}{additional arguments}

\item{ind_vars}{a character vector naming the columns to extract factor levels from}
}
\value{
a dataframe of class "factor.grid" for passing to additional factorial-design functions
}
\description{
This the generic create_fgrid function.

A vectorized vesion of the \code{create_fgrid} function

Create a factor grid object from a named list of factors

Create a \code{factor.grid} df from a dataframe
}
\seealso{
Other factorial-design functions: \code{\link{aov_fgrid}},
  \code{\link{combine_factor_levels}},
  \code{\link{expand_fgrid}}, \code{\link{factor_summary}},
  \code{\link{summary_form_grid}}
}
