% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_grid-function.R
\name{df_to_factgrid}
\alias{df_to_factgrid}
\title{Create a \code{factor.grid} df from a data set}
\usage{
df_to_factgrid(data, ind_vars)
}
\arguments{
\item{data}{data that will be converted to a \code{factor.grid} dataframe}

\item{ind_vars}{a character vector naming the columns to extract factor levels from}
}
\description{
Create a \code{factor.grid} df from a data set
}
\examples{

mydata <- data.frame(
factorA = rep(c("sugar","control"), each = 2),
factorB = rep(c("coffee","tea"), times = 2),
score  = runif(n = 4,min = 10,max = 100)
)

df_to_factgrid(mydata, c("factorA","factorB"),
}
\seealso{
Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_filter}},
  \code{\link{factor_grid.list}}, \code{\link{factor_grid}}
}
