% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_maker-function.R
\name{dots_maker}
\alias{dots_maker}
\title{Helper functions for sweet_dot}
\usage{
dots_maker(.data, ...)
}
\arguments{
\item{.data}{data to generate dots from}

\item{...}{optional arguments passed to/from other methods (ignore)}
}
\value{
a formula
}
\description{
Helper functions for sweet_dot
}
\seealso{
Other SE dplyr convenience functions: \code{\link{select_dots}},
  \code{\link{sweet_dots}}
}
