% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_fgrid-function.R
\name{expand_fgrid}
\alias{expand_fgrid}
\alias{expand_fgrid.default}
\alias{expand_fgrid.factor.grid}
\title{Data-filtering conditions as vectors}
\usage{
expand_fgrid(.data, ...)

\method{expand_fgrid}{default}(.data, ...)

\method{expand_fgrid}{factor.grid}(.data, ...)
}
\arguments{
\item{.data}{a \code{factor.grid} dataframe generated by  or}

\item{...}{additional arguments}
}
\value{
a data.frame of factorial-based filtering condition or a  list for passing to additional
}
\description{
Creates a factor grid dataframe of data-filtering conditions for passing to additional to additional factorial-design functions.
}
\seealso{
Other factorial-design functions: \code{\link{aov_fgrid}},
  \code{\link{combine_factor_levels}},
  \code{\link{create_fgrid}}, \code{\link{factor_summary}},
  \code{\link{summary_form_grid}}
}
