% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_fgrid_helpers.R
\name{grow_fgrid}
\alias{grow_fgrid}
\alias{drop_fgrid_levels}
\alias{number_of_levels}
\alias{label_levels}
\alias{filter_from_levels}
\title{Helper functions for creating factor_filter_columns}
\usage{
grow_fgrid(.ind_var, .data, ...)

drop_fgrid_levels(.data, ...)

number_of_levels(.data, ...)

label_levels(.data, ...)

filter_from_levels(.data, ...)
}
\arguments{
\item{.ind_var}{one ore more independent variables}

\item{.data}{a df.factor.grid, or part of the grid}

\item{...}{additional arguments (ignore)}
}
\value{
a data.frame of factorial-based filtering condition or a  list for passing to additional

drop levels

a vector of length of .data

a vector of labels representing crossed-factor levels

a vector identifying levels for filtering a dataframe
}
\description{
Helper functions for creating factor_filter_columns
}
