% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_grid-function.R
\name{extract_factor_grid}
\alias{extract_factor_grid}
\title{Create a factor grid object from a data set}
\usage{
extract_factor_grid(data, ind_vars)
}
\arguments{
\item{data}{data that will be converted to a dataframe of the "factor.grid" class}

\item{ind_vars}{a character vector naming the columns to extract factor levels from}
}
\description{
Create a factor grid object from a data set
}
\examples{

mydata <- data.frame(
factorA = rep(c("sugar","control"), each = 2),
factorB = rep(c("coffee","tea"), times = 2),
score  = runif(n = 4,min = 10,max = 100)
)

extract_factor_grid(mydata, c("factorA","factorB"),

}
\seealso{
Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{factorial_grid.list}},
  \code{\link{factorial_grid}}
}
