% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOffactor_filter-function.R,
%   R/factor_filter-function.R
\name{factor_filter.factor.grid}
\alias{factor_filter.factor.grid}
\alias{factor_filter.factor.grid}
\title{Derive filtering conditions from a factor grid}
\usage{
\method{factor_filter}{factor.grid}(data, trunc_label = TRUE,
  for_aov = TRUE, ...)

\method{factor_filter}{factor.grid}(data, trunc_label = TRUE,
  for_aov = TRUE, ...)
}
\arguments{
\item{trunc_label}{Should the output contain a column with printing/reading-friendly, truncuated labels to indicate which factor(s)/levels(s) are being used to filter? Default is TRUE. Mutlple levels of a single factor are separated with a colon: (e.g., <code>fact1level1:fact1level2</code>); factors are separated by a double-colon: (e.g., <code>fact1level1:fact1level2::fact2level1</code>)}

\item{for_aov}{A logical indicating whether the data be returned for passing to a function in @family aov-out functions? Defaults to TRUE.}

\item{trunc_label}{Should the output contain a column with printing/reading-friendly, truncuated labels to indicate which factor(s)/levels(s) are being used to filter? Default is TRUE. Mutlple levels of a single factor are separated with a colon: (e.g., <code>fact1level1:fact1level2</code>); factors are separated by a double-colon: (e.g., <code>fact1level1:fact1level2::fact2level1</code>)}

\item{for_aov}{A logical indicating whether the data be returned for passing to a function in @family aov-out functions? Defaults to TRUE.}
}
\description{
Derive filtering conditions from a factor grid

Derive filtering conditions from a factor grid
}
\examples{

my_filtertgrid <- factor_filter(my_factgrid)
)

my_filtertgrid <- factor_filter(my_factgrid)
)
}
\seealso{
Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter}},
  \code{\link{factor_grid.list}}, \code{\link{factor_grid}}

Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter}},
  \code{\link{factor_grid.list}}, \code{\link{factor_grid}}
}
