% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOffactor_filter-function.R,
%   R/factor_filter-function.R
\name{factor_filter}
\alias{factor_filter}
\alias{factor_filter.default}
\alias{factor_filter}
\alias{factor_filter.default}
\title{Data-filtering conditions as vectors}
\usage{
factor_filter(data, ...)

\method{factor_filter}{default}(data, ...)

factor_filter(data, ...)

\method{factor_filter}{default}(data, ...)
}
\arguments{
\item{data}{a \code{factor.grid} dataframe generated by \code{\link{factor_grid}} or \code{\link{df_to_factgrid}}}

\item{...}{further arguments passed to or from other methods}

\item{data}{a \code{factor.grid} dataframe generated by \code{\link{factor_grid}} or \code{\link{df_to_factgrid}}}

\item{...}{further arguments passed to or from other methods}
}
\value{
a data.frame of factorial-based filtering condition or a \code{filter.grid} list for passing to additional \link{factorial-design functions}

a data.frame of factorial-based filtering condition or a \code{filter.grid} list for passing to additional \link{factorial-design functions}
}
\description{
Creates a  \code{factor.grid} dataframe of data-filtering conditions for passing to additional to additional \link{factorial-design functions}.

A vectorized vesion of the \code{factor_filter} function

Creates a  \code{factor.grid} dataframe of data-filtering conditions for passing to additional to additional \link{factorial-design functions}.

A vectorized vesion of the \code{factor_filter} function
}
\details{
Formula-strings for filtering a data set based on the levels of crossed factors

Formula-strings for filtering a data set based on the levels of crossed factors
}
\seealso{
Other factorial-design functions: \code{\link{aov_forms}},
  \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_grid.list}},
  \code{\link{factor_grid}}, \code{\link{factor_summary}}

Other factorial-design functions: \code{\link{aov_forms}},
  \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_grid.list}},
  \code{\link{factor_grid}}, \code{\link{factor_summary}}

Other factorial-design functions: \code{\link{aov_forms}},
  \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_grid.list}},
  \code{\link{factor_grid}}, \code{\link{factor_summary}}

Other factorial-design functions: \code{\link{aov_forms}},
  \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_grid.list}},
  \code{\link{factor_grid}}, \code{\link{factor_summary}}
}
