% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_filter_helpers-function.R
\name{ff_forms_col}
\alias{ff_forms_col}
\alias{ff_number_levels_col}
\alias{ff_label_col}
\alias{ff_filter_f_col}
\title{Helper functions for creating factor_filter_columns}
\usage{
ff_forms_col(.data, select_column, ...)

ff_number_levels_col(.data, select_column, ...)

ff_label_col(.data, select_column, ..., match_pattern, replace_pattern)

ff_filter_f_col(.data, select_column, ...)
}
\arguments{
\item{.data}{a df.factor.grid, or part of the grid}

\item{select_column}{the name of the column used to generate the ff_ column needed}

\item{...}{additional arguments (ignore)}

\item{match_pattern}{a pattern to match for breaking label vectors apart=(Default is "_nm")}

\item{replace_pattern}{a pattern to use for replacing matches}
}
\value{
a vector of filtering formulas

for <code>ff_number_levels_column</code>, a vector of length of .data

for <code>ff_label_column</code>, a vector of filtering formulas

for <code>ff_number_levels_column</code>, a vector of length of .data
}
\description{
Helper functions for creating factor_filter_columns

Create a "number_levels_column"

Create a ff_label_column

Create a ff_filter_f_column
}
