% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_grid-function.R
\name{factor_grid.list}
\alias{factor_grid.list}
\title{Create a factor grid object from a named list of factors}
\usage{
\method{factor_grid}{list}(data, ...)
}
\description{
Create a factor grid object from a named list of factors
}
\examples{
# create a factorial grid data frame by passing a named list of factors/levels
factor_grid(list("myvar1" = c("A","B","C","D"),
"myvar2" = c("long","short")))
}
\seealso{
Other factorial-design functions: \code{\link{aov_forms}},
  \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_filter}}, \code{\link{factor_grid}},
  \code{\link{factor_summary}}
}
