% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_grid-function.R
\name{factor_grid}
\alias{factor_grid}
\alias{factor_grid.default}
\title{Create a data frame of variable combinations for a fully-crossed factorial design}
\usage{
factor_grid(data, ...)

\method{factor_grid}{default}(data, ...)
}
\arguments{
\item{data}{data that will be converted to a \code{factor.grid} dataframe}

\item{...}{further arguments passed to or from other methods}
}
\value{
a dataframe of class "factor.grid" for passing to additional \link{factorial-design functions}
}
\description{
This the generic factor_grid function. See Details.

A vectorized vesion of the \code{factor_grid} function
}
\details{
\itemize{
  \item \code{\link{factor_grid.list}} for a named list
  \item \code{\link{df_to_factgrid}} for a dataframe
}
}
\seealso{
Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_filter}},
  \code{\link{factor_grid.list}}

Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_filter}},
  \code{\link{factor_grid.list}}
}
