% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_summary-function.R
\name{factor_summary}
\alias{factor_summary}
\title{Summary stats for factors}
\usage{
factor_summary(data, dv, iv, grp_factor)
}
\arguments{
\item{data}{to summarize}

\item{dv}{the dependent variable}

\item{iv}{a vector wiht the names of the independent variables}

\item{grp_factor}{a grouping factor}
}
\value{
a data.frame of summary stats
}
\description{
Summary stats for factors
}
\seealso{
Other factorial-design functions: \code{\link{aov_forms}},
  \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_filter}},
  \code{\link{factor_grid.list}}, \code{\link{factor_grid}}
}
