% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_summary_helpers-function.R
\name{summary_subset}
\alias{summary_subset}
\alias{summary_combine}
\title{Helper functions for creating factor_summary_data}
\usage{
summary_subset(data, cols, meas, ...)

summary_combine(data, ...)
}
\arguments{
\item{data}{for which a summary is to be computed}

\item{cols}{FALSE (default) for collapsing over all variables, or a vector of variables to collapses over in stages}

\item{meas}{dependent measure to compute summaries over}

\item{...}{ignored arguments to other functions}
}
\description{
Helper functions for creating factor_summary_data

Combine summaries
}
