% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_grid-function.R
\name{factorial_grid.list}
\alias{factorial_grid.list}
\title{Create a factor grid object from a named list of factors}
\usage{
\method{factorial_grid}{list}(data, ...)
}
\description{
Create a factor grid object from a named list of factors
}
\examples{
# create a factorial grid data frame by passing a named list of factors/levels
factorial_grid(list("myvar1" = c("A","B","C","D"),
"myvar2" = c("long","short")))
}
\seealso{
Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{extract_factor_grid}},
  \code{\link{factorial_grid}}
}
