% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_grid-function.R
\name{factorial_grid}
\alias{factorial_grid}
\alias{factorial_grid.default}
\title{Create a data frame of variable combinations for a fully-crossed factorial design}
\usage{
factorial_grid(data, ...)

\method{factorial_grid}{default}(data, ...)
}
\arguments{
\item{data}{data that will be converted to a dataframe of the "factor.grid" class}

\item{...}{further arguments passed to or from other methods}
}
\value{
a dataframe of class "factor.grid" for passing to additional functions
}
\description{
This the generic factorial_grid function. See the following functions
for the details about different data structures:

A vectorized vesion of the \code{factorial_grid} function
}
\details{
\itemize{
  \item \code{\link{factorial_grid.list}} for a named list
  \item \code{\link{factorial_grid.data.frame}} for a dataframe
}
}
\seealso{
Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{extract_factor_grid}},
  \code{\link{factorial_grid.list}}

Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{extract_factor_grid}},
  \code{\link{factorial_grid.list}}
}
