% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist_to_dots-function.R
\name{flist_to_dots}
\alias{flist_to_dots}
\title{Return a list of formulas or a character vector of such a list for the NSE dots arugment.}
\usage{
flist_to_dots(formlist)
}
\arguments{
\item{formlist}{Either a list of comma-separated RHS formulas, as 'list(~f1, ~f2)', or a character vector of such a list: \\"list(~col1,col2)\\".}
}
\value{
An expression that can be interpred as a \code{dots} argument.
}
\description{
\code{flist_to_dots} prepares a list of formulas or a character vector of such a list to be passed to the \code{.dots} argument of the NSE versions of \code{\link[dplyr]{index}} functions like \code{filter_}.
}

