% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_column-functions.R
\name{iv_between_col}
\alias{iv_between_col}
\alias{aov_vars_col}
\alias{aov_select_col}
\title{Helper functions for aov_forms}
\usage{
iv_between_col(data, btw_var)

aov_vars_col(data, ...)

aov_select_col(data, s_dep_var, s_grp_factor)
}
\arguments{
\item{data}{data passed from aov_forms}

\item{btw_var}{a between-subjects variable}

\item{...}{data to pass to or from other functions}

\item{s_dep_var}{dependent variable selector}

\item{s_grp_factor}{name of the grouping factor}
}
\value{
an aov-ready column
}
\description{
Helper functions for aov_forms

Anova variables column

Anova select column
}
