% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range-function.R
\name{get_range}
\alias{get_range}
\title{Return p-value as a range from a number.}
\usage{
get_range(p)
}
\arguments{
\item{p}{A number for which a p-value rage should be returned.}
}
\value{
The significance range: e.g., "<.001"
}
\description{
For when p-value stars are not appropriate, etc.
}
