% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range.tex-function.R
\name{get_range.tex}
\alias{get_range.tex}
\title{Return p-value as a range from a number without greater-than or less-than sign.}
\usage{
get_range.tex(p)
}
\arguments{
\item{p}{A number for which a p-value rage should be returned.}
}
\value{
The bare significance range: e.g., "001"
}
\description{
May be useful when numbers will be used in a LaTex document or places where only the numerical value as a character is wanted.
}

