% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stars-function.R
\name{get_stars}
\alias{get_stars}
\title{Return p-value significance star(s) from a number.}
\usage{
get_stars(p)
}
\arguments{
\item{p}{A number for which a p-value star should be returned.}
}
\value{
A character star: e.g., ***,**,*, ., NULL
}
\description{
Uses the standard range of intervals to determine the nubmer of stars.
}

