% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_x-function.R
\name{labelled_x}
\alias{labelled_x}
\title{A wrapper for \code{\link[sjmisc]{get_label}} suitable for use with \code{ggplot2}'s \code{xlab} function.}
\usage{
labelled_x(xdata, xname, tex = TRUE)
}
\arguments{
\item{xdata}{A dataframe with labelled columns.}

\item{xname}{The (bare) name of the column to be used as the x label.}

\item{tex}{Logical (default = TRUE); Does the label contain LaTeX?}
}
\value{
If tex = TRUE, the column label as a \code{\link[grDevices]{plotmath}} expression (see \code{\link[latex2exp]{TeX}} for details; or a character string.
}
\description{
For dataframes with labels (see \code{\link[sjmisc]{get_label}}),  \code{labelled_x} returns an epxression (for LaTex = TRUE) or string that can be passed to \code{xlab} in a \code{ggplot} (see \code{\link[ggplot2]{ggtitle}}).
}

