% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_y-function.R
\name{labelled_y}
\alias{labelled_y}
\title{A wrapper for \code{\link[sjmisc]{get_label}} suitable for use with \code{ggplot2}'s \code{ylab} function.}
\usage{
labelled_y(ydata, yname, tex = TRUE)
}
\arguments{
\item{ydata}{A dataframe with labelled columns.}

\item{yname}{The (bare) name of the column to be used as the y label.}

\item{tex}{Logical (default = TRUE); Does the label contain LaTeX?}
}
\value{
If tex = TRUE, the column label as a \code{\link[grDevices]{plotmath}} expression (see \code{\link[latex2exp]{TeX}} for details; or a character string.
}
\description{
For dataframes with labels (see \code{\link[sjmisc]{get_label}}),  \code{labelled_y} returns an epxression (for LaTex = TRUE) or string that can be passed to \code{ylab} in a \code{ggplot} (see \code{\link[ggplot2]{ggtitle}}).
}

