% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noz-function.R
\name{noz}
\alias{noz}
\title{Remove leading zeros from a number.}
\usage{
noz(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
\code{x} as a character with leading zeroes removed
}
\description{
Useful when numbers are going to be used in APA-foramatted documents. Note that this concerts the number to a character vector and is not guaranteed to produce the appropriate result in all cases.
}
