% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_stars-function.R
\name{p_stars}
\alias{p_stars}
\title{Star-type signigicance symbols}
\usage{
p_stars(p, show_ns = TRUE)
}
\arguments{
\item{p}{value whose significance is to be indicated with a star}

\item{show_ns}{should non-significant values be shown with "n.s."? Default is TRUE}
}
\value{
text stars
}
\description{
Star-type signigicance symbols
}
