% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pround-function.R
\name{pround}
\alias{pround}
\title{Helper function for sweetp}
\usage{
pround(p)
}
\arguments{
\item{p}{a numeric or character p-value}

\item{...}{further arguments passed to or from other methods}
}
\value{
p as APA-rounded character vector
}
\description{
Helper function for sweetp
}
\seealso{
Other sweetstat functions: \code{\link{statround}},
  \code{\link{sweetp}}, \code{\link{sweetstat}}
}
