% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_dir-function.R
\name{root_dir}
\alias{root_dir}
\title{Shortcut for specifying paths relative to root.}
\usage{
root_dir(pathpart)
}
\arguments{
\item{pathpart}{A string to the path}
}
\value{
A character vector with the path to the root directory appended to the path-part specified.
}
\description{
\code{root_dir} is a wrapper around \code{getwd} that appends the path-part to the location specified by the user-supplied argument.
}
\examples{

# Supply path-part without preceding forward slash

my_path <-
 root_dir("myfolder/test.txt")

 my_path
}
