% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter-function.R
\name{scatter}
\alias{scatter}
\title{A wrapper for a \code{ggplot2} scatterplot with a custom minimal theme.}
\usage{
scatter(df, x, y, line_col = "#5B1A18", fill_col = "#FD6467",
  facet_form = NULL, grp_by = NULL, add_stats = TRUE)
}
\arguments{
\item{df}{The dataframe being plotted.}

\item{x, y}{The (bare) names of the column to be used as the \code{x} and \code{y} in the scatterplot .}

\item{line_col}{The color of the \code{\link[ggplot2]{geom_smooth}} line; default is "#5B1A18" from the "GrandBudapest" palette in \link[https://github.com/karthik/wesanderson]{wes_palettes}.}

\item{fill_col}{The color of the \code{\link[ggplot2]{geom_smooth}} fill; default is "#FD6467" from the "GrandBudapest" palette in \link[https://github.com/karthik/wesanderson]{wes_palettes}.}

\item{facet_form}{A quoted string to be used as formula for facetting the figures, as, for example "factor1+factor2"; default = \code{NULL}.}

\item{grp_by}{If the scatter plot will include faceting, a list of RHS formulas (see \code{\link[lazyeval]{f_rhs}}) to be passed to \code{\link[dissertate]{scatter_stats}}; default = \code{NULL}. Ensures stats are calculated separately for each group.}

\item{add_stats}{Logical; Add model fit info to the figure? Default = \code{TRUE}.}
}
\value{
A ggplot2 object.
}
\description{
\code{scatter} streamlines the process of creating a scatterplot in \code{ggplot2}, with some flexibilty for faceting (see \code{\link[ggplot2]{ggplot2}}).
}

