% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_stats-function.R
\name{scatter_stats}
\alias{scatter_stats}
\title{Compute useful statistics for use with \code{\link[dissertate]{scatter}}.}
\usage{
scatter_stats(statdf, statx, staty, grp_by_form = NULL)
}
\arguments{
\item{statdf}{The dataframe being plotted.}

\item{statx, staty}{The (bare) names of the column in \code{statdf} to be used as the \code{x} and \code{y} in the scatterplot.}

\item{grp_by_form}{If the scatter plot will include faceting, a list of RHS formulas (see \code{\link[lazyeval]{f_rhs}}) to group \code{statdf} by; default = \code{NULL}. Ensures stats are calculated separately for each group.}
}
\value{
A dataframe with model fit statistics and minimum and maximum values for \code{statx} and \code{staty}
}
\description{
Provides statistics that are used in \code{\link[dissertate]{scatter}} and passed to \code{\link[ggplot2]{ggplot2}}) to display model fit results and determin appropriate plot scales, etc.
}
