% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se-function.R
\name{se}
\alias{se}
\title{Compute the standard error of \code{x}.}
\usage{
se(x)
}
\arguments{
\item{x}{A numeric vector (such as the column of a dataframe).}
}
\value{
The standard error of \code{x}
}
\description{
Standard error.
}
