% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.data.frame-function.R
\name{sort.data.frame}
\alias{sort.data.frame}
\title{Sort a data frame by column}
\usage{
\method{sort}{data.frame}(x, decreasing = FALSE, by = 1, ...)
}
\arguments{
\item{x}{a data frame}

\item{decreasing}{Decreasing oder?}

\item{by}{index to sort on}

\item{...}{further arguments passed to or from other methods (ignored)}
}
\value{
a sorted data frame
}
\description{
Creates a table of aov summary
}
\examples{
##sort(iris, by="Sepal.Length")
##sort(iris, by=c("Species","Sepal.Length"))
##sort(iris, by=1:2)
##sort(iris, by="Sepal.Length",decreasing=TRUE)
}
