% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error-function.R
\name{standard_error}
\alias{standard_error}
\title{Compute a standard error statistic for a vector}
\usage{
standard_error(x)
}
\arguments{
\item{x}{a vector of numeric values}
}
\value{
a standard error
}
\description{
Compute a standard error statistic for a vector
}
