% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statround-function.R
\name{statround}
\alias{statround}
\title{Helper function for sweetstat}
\usage{
statround(s)
}
\arguments{
\item{s}{a number to be rounded}

\item{...}{further arguments passed to or from other methods}
}
\value{
a rounded numeric
}
\description{
Helper function for sweetstat
}
\seealso{
Other sweetstat functions: \code{\link{pround}},
  \code{\link{sweetp}}, \code{\link{sweetstat}}
}
