% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_form_grid-function.R
\name{summary_form_grid}
\alias{summary_form_grid}
\alias{summary_form_grid_generate}
\alias{summary_form_grid.default}
\alias{summary_form_grid_vectorized}
\title{Create sumamry formulas from a filter.grid}
\usage{
summary_form_grid(.data, ..., .dep_var, .grp_var)

summary_form_grid_generate(.data, ..., .dep_var, .grp_var)

summary_form_grid.default(.data, .dep_var, .grp_var, ...)

summary_form_grid_vectorized(.data, ..., .dep_var, .grp_var)
}
\arguments{
\item{.data}{a dataframe or a named list of data frames}

\item{...}{further arguments passed to or from other methods}

\item{.dep_var}{a vector naming the depend. var.}

\item{.grp_var}{an optional grouping factor}
}
\value{
a data.frame with a  class attribute for passing to additional
}
\description{
Creates a dataframe of formulas for summary from a filter grid
}
\seealso{
Other factorial-design functions: \code{\link{aov_form_grid}},
  \code{\link{combine_factor_levels}},
  \code{\link{factor_filter}}, \code{\link{factor_grid}},
  \code{\link{factor_summary}}
}
