% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_form_helpers-function.R
\name{summary_vars_col}
\alias{summary_vars_col}
\alias{summary_select_col}
\alias{summary_index_col}
\alias{summary_clean_cols}
\title{Summary form helpers}
\usage{
summary_vars_col(.data, ...)

summary_select_col(.data, .dep_var, .grp_var, ...)

summary_index_col(.data, .dep_var, .grp_var, ...)

summary_clean_cols(.data, ...)
}
\arguments{
\item{.data}{a  .dataframe}

\item{...}{further arguments passed to or from other methods}

\item{.dep_var}{a vector naming the depend. var.}

\item{.grp_var}{the grouping factor}
}
\value{
a data.frame with a  class attribute for passing to additional
}
\description{
Summary form helpers

Anova select column
}
