% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_run-function.R
\name{summary_run}
\alias{summary_run}
\title{Create AOV formulas from a filter.grid}
\usage{
summary_run(.data, .spec, ...)
}
\arguments{
\item{.data}{a data frame or list of named data frames containing source data}

\item{.spec}{a summary grid dataframe}

\item{...}{passed to/from other functions (ignored)}
}
\value{
a data.frame with summary stats
}
\description{
Creates a dataframe of formulas for aovs from a filter grid
}
