% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_dots-function.R
\name{sweet_dots}
\alias{sweet_dots}
\alias{sweet_dots.character}
\alias{sweet_dots.integer}
\alias{sweet_dots.formula}
\alias{sweet_dots.default}
\title{Create dots}
\usage{
sweet_dots(.data, ...)

\method{sweet_dots}{character}(.data, ...)

\method{sweet_dots}{integer}(.data, ...)

\method{sweet_dots}{formula}(.data, ...)

\method{sweet_dots}{default}(.data, ...)
}
\arguments{
\item{.data}{data to pass to selects "dots" parameter}

\item{...}{optional arguments passed to/from other methods}
}
\value{
a formula
}
\description{
The function takes a string, a vector, a formula, a list of formulas, or a string representing such a list and "quotes" it for passing to the dplyr functions.
}
\examples{

# Pass a string
 dplyr::select_(iris, .dots = sweet_dots("Sepal.Length"))
 dplyr::select_(iris, .dots = sweet_dots("dplyr::matches('Sepal')"))
 dplyr::filter_(iris, .dots = sweet_dots("Sepal.Length == 5.1 & Species != 'setosa'"))

# Pass a vector
dplyr::select_(iris, .dots = sweet_dots(c("Sepal.Length","Species")))
}
\seealso{
Other SE dplyr convenience functions: \code{\link{dots_maker}},
  \code{\link{select_dots}}
}
