% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_p-function.R
\name{sweet_p}
\alias{sweet_p}
\alias{sweet_p.numeric}
\alias{sweet_p.character}
\alias{sweet_p.integer}
\alias{sweet_p.default}
\title{APA-style decimal rounding for p-values}
\usage{
sweet_p(.stat, ...)

\method{sweet_p}{numeric}(.stat, ...)

\method{sweet_p}{character}(.stat, ...)

\method{sweet_p}{integer}(.stat, ...)

\method{sweet_p}{default}(.stat, ...)
}
\arguments{
\item{.stat}{a numeric value representing a statistic}

\item{...}{(ignored)}
}
\value{
stat as APA-rounded character vector
}
\description{
This is the generic version of the function.

Pass a numeric  value to sweet_p

Pass a numeric string value to sweet_p

Pass a integer  value to sweet_p

Default method for sweetening a statistic
}
\seealso{
Other sweet_stat functions: \code{\link{sweet_stat}}
}
