% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_stat-function.R
\name{sweet_stat}
\alias{sweet_stat}
\alias{sweet_stat.numeric}
\alias{sweet_stat.character}
\alias{sweet_stat.integer}
\alias{sweet_stat.default}
\title{APA-style decimal rounding for stats}
\usage{
sweet_stat(.stat, ...)

\method{sweet_stat}{numeric}(.stat, ...)

\method{sweet_stat}{character}(.stat, ...)

\method{sweet_stat}{integer}(.stat, ...)

\method{sweet_stat}{default}(.stat, ...)
}
\arguments{
\item{.stat}{a numeric value representing a statistic}

\item{...}{pass "rnd_digit" to set digits (defaults to 2)}
}
\value{
stat as APA-rounded character vector
}
\description{
This is the generic version of the function.

Pass a numeric  value to sweet_stat

Pass a numeric string value to sweet_stat

Pass a integer  value to sweet_stat

Default method for sweetening a statistic
}
\seealso{
Other sweet_stat functions: \code{\link{sweet_p}}
}
