% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_tidy-function.R
\name{sweet_tidy}
\alias{sweet_tidy}
\title{Modified version of tidy}
\usage{
sweet_tidy(.data, ...)
}
\arguments{
\item{.data}{an aovlist object}

\item{...}{further arguments passed to or from other methods}
}
\value{
a data.frame with the results of an anova
}
\description{
Modified version of tidy
}
