% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetp-function.R
\name{sweetp}
\alias{sweetp}
\alias{sweetp.numeric}
\alias{sweetp.character}
\alias{sweetp.integer}
\alias{sweetp.default}
\title{APA-style decimal rounding for p-values}
\usage{
sweetp(.stat, ...)

\method{sweetp}{numeric}(.stat, ...)

\method{sweetp}{character}(.stat, ...)

\method{sweetp}{integer}(.stat, ...)

\method{sweetp}{default}(.stat, ...)
}
\arguments{
\item{.stat}{a numeric value representing a statistic}

\item{...}{(ignored)}
}
\value{
stat as APA-rounded character vector
}
\description{
This is the generic version of the function.

Pass a numeric  value to sweetp

Pass a numeric string value to sweetp

Pass a integer  value to sweetp

Default method for sweetening a statistic
}
\seealso{
Other sweetstat functions: \code{\link{pround}},
  \code{\link{statround}}, \code{\link{sweetpround}},
  \code{\link{sweetround}}, \code{\link{sweetstat}}
}
