% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetstat-function.R
\name{sweetstat}
\alias{sweetstat}
\alias{sweetstat.numeric}
\alias{sweetstat.character}
\alias{sweetstat.integer}
\alias{sweetstat.default}
\title{APA-style decimal rounding for stats}
\usage{
sweetstat(.stat, ...)

\method{sweetstat}{numeric}(.stat, ...)

\method{sweetstat}{character}(.stat, ...)

\method{sweetstat}{integer}(.stat, ...)

\method{sweetstat}{default}(.stat, ...)
}
\arguments{
\item{.stat}{a numeric value representing a statistic}

\item{...}{pass "rnd_digit" to set digits (defaults to 2)}
}
\value{
stat as APA-rounded character vector
}
\description{
This is the generic version of the function.

Pass a numeric  value to sweetstat

Pass a numeric string value to sweetstat

Pass a integer  value to sweetstat

Default method for sweetening a statistic
}
\seealso{
Other sweetstat functions: \code{\link{pround}},
  \code{\link{statround}}, \code{\link{sweetround}}
}
