% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_form_grid-function.R
\name{aov_form_grid}
\alias{aov_form_grid}
\alias{aov_form_grid_generate}
\alias{aov_form_grid.default}
\title{Create AOV formulas from a filter.grid}
\usage{
aov_form_grid(.data, ...)

aov_form_grid_generate(.data, ..., .dep_var, .grp_var, .btw_var = FALSE)

\method{aov_form_grid}{default}(.data, .dep_var, .grp_var, .btw_var = FALSE,
  ...)
}
\arguments{
\item{.data}{a  .dataframe}

\item{...}{further arguments passed to or from other methods}

\item{.dep_var}{a vector naming the depend. var.}

\item{.grp_var}{an optional grouping factor}

\item{.btw_var}{an optional vector naming the between-subjects variable}
}
\value{
a data.frame with a  class attribute for passing to additional
}
\description{
Creates a dataframe of formulas for aovs from a filter grid

Create ANOVA formulas from a filter.grid data frame

Create ANOVA formulas from a filter.grid data frame
}
\seealso{
Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{factor_filter}}, \code{\link{factor_grid}},
  \code{\link{factor_summary}}
}
