% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_fgrid_helpers.R
\name{aov_formulate}
\alias{aov_formulate}
\alias{aov_terms_col}
\alias{iv_between_col}
\alias{aov_vars_col}
\alias{aov_select_col}
\alias{aov_groupby_col}
\alias{aov_index_col}
\alias{aov_clean_cols}
\title{AOV form helpers}
\usage{
aov_formulate(.data, .dep_var, .grp_var, ...)

aov_terms_col(.data, ...)

iv_between_col(.data, .btw_var, ...)

aov_vars_col(.data, ...)

aov_select_col(.data, .grp_var, ...)

aov_groupby_col(.data, .grp_var, ...)

aov_index_col(.data, .dep_var, .grp_var, ...)

aov_clean_cols(.data, ...)
}
\arguments{
\item{.data}{a  .dataframe}

\item{.dep_var}{a vector naming the depend. var.}

\item{.grp_var}{the grouping factor}

\item{...}{further arguments passed to or from other methods}

\item{.btw_var}{an optional between-subjects variable}
}
\value{
a data.frame with a  class attribute for passing to additional
}
\description{
AOV form helpers

Between-subjects variable column

Anova variables column

Anova select column

Anova aggregrate dv group-by  column
}
