% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_form_helpers-function.R
\name{aov_formulate}
\alias{aov_formulate}
\alias{aov_formulate.default}
\alias{aov_formulate.between}
\alias{aov_formulate.within}
\alias{aov_terms_col}
\alias{iv_between_col}
\alias{aov_vars_col}
\alias{aov_select_col}
\title{AOV form helpers}
\usage{
aov_formulate(.data, ...)

\method{aov_formulate}{default}(.data, .dep_var, .grp_var, ...)

\method{aov_formulate}{between}(.data, ..., .dep_var, .grp_var, .btw_var)

\method{aov_formulate}{within}(.data, ..., .dep_var, .grp_var)

aov_terms_col(.data, .btw_var = FALSE, ...)

iv_between_col(.data, .btw_var, ...)

aov_vars_col(.data, .rm_col = FALSE, ...)

aov_select_col(.data, .dep_var, .grp_var, ...)
}
\arguments{
\item{.data}{a  .dataframe}

\item{...}{further arguments passed to or from other methods}

\item{.dep_var}{a vector naming the depend. var.}

\item{.grp_var}{the grouping factor}

\item{.btw_var}{an optional vector naming the between-subjects variable}

\item{.rm_col}{an optional vector naming columns to remove}
}
\value{
a data.frame with a  class attribute for passing to additional
}
\description{
AOV form helpers

Between-subjects variable column

Anova variables column

Anova select column
}
