% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfaov_forms-function.R
\name{aov_forms}
\alias{aov_forms}
\title{Create AOV formulas from a filter.grid}
\usage{
aov_forms(data, ...)
}
\arguments{
\item{data}{a \code{filter.grid} dataframe}

\item{...}{further arguments passed to or from other methods}
}
\value{
a data.frame with a \code{filter.grid} class attribute for passing to additional \link{factorial-design functions}
}
\seealso{
Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_filter}},
  \code{\link{factor_grid.list}}, \code{\link{factor_grid}}
}
