% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_run-function.R
\name{aggregrate_data}
\alias{aggregrate_data}
\alias{aov_run}
\title{Create AOV formulas from a filter.grid}
\usage{
aggregrate_data(.data, .spec, ...)

aov_run(.data, .spec, ...)
}
\arguments{
\item{.data}{data to be analysed}

\item{.spec}{specs for anovas}

\item{...}{further arguments passed to or from other methods}
}
\value{
a data.frame with the results of an anova
}
\description{
Creates a dataframe of formulas for aovs from a filter grid
}
