% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA_txt_num.R
\name{APA_txt_num}
\alias{APA_txt_num}
\title{APA numbers-as-text}
\usage{
APA_txt_num(x, ...)
}
\arguments{
\item{x}{a  .dataframe}

\item{...}{further arguments passed to or from other methods}
}
\value{
a number expressed as text
}
\description{
APA numbers-as-text
}
