% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_resid_function.R
\name{append_resids}
\alias{append_resids}
\title{Append Residuals Function}
\usage{
append_resids(dataframe, id.group, lm.form)
}
\arguments{
\item{dataframe}{a dataframe with at least two numeric/integer columns suitable for fitting to `stats::lm(y ~ x)`}

\item{id.group}{a list of one or more RHS formulas naming the dataframe column(s) that will be used to create a unique row identifier for binding the column of residual values: list(~col1,~col2); if NULL (default), assumes each row is unique )}

\item{lm.form}{= a list of exactly two RHS formulas naming the dataframe columns over which residulas will be computed, or else a formula to be passed to `stats::lm()`}
}
\value{
a copy of the original dataframe with residual columns appended
}
\description{
Append Residuals Function
}
\keyword{residuals}

