% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_factor_levels-function.R
\name{combine_factor_levels}
\alias{combine_factor_levels}
\title{A vectorized version of \code{utils::combn}}
\usage{
combine_factor_levels(x, m, FUNV = NULL, simplify, byrow)
}
\arguments{
\item{x}{vector source for combinations, or integer \code{n} for
    \code{x <- \link{seq_len}(n)}.}

\item{m}{number of elements to choose.}

\item{simplify}{logical indicating if the result should be simplified
    to an \code{\link{array}} (typically a \code{\link{matrix}}); if
    FALSE, the function returns a \code{\link{list}}.  Note that when
    \code{simplify = TRUE} as by default, the dimension of the result is
    simply determined from \code{FUN(\var{1st combination})} (for
    efficiency reasons).  This will badly fail if \code{FUN(u)} is not of
    constant length.}
}
\description{
A vectorized version of \code{utils::combn}
}
\seealso{
Other factorial-design functions: \code{\link{df_to_factgrid}},
  \code{\link{factor_filter.factor.grid}},
  \code{\link{factor_filter}},
  \code{\link{factor_grid.list}}, \code{\link{factor_grid}}
}
