% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_filter-function.R
\name{factor_filter}
\alias{factor_filter}
\alias{factor_filter.default}
\alias{factor_filter.factor.grid}
\title{Data-filtering conditions as vectors}
\usage{
factor_filter(data)

\method{factor_filter}{default}(data)

\method{factor_filter}{factor.grid}(data, trunc_label = TRUE,
  for_aov = TRUE)
}
\arguments{
\item{data}{a \code{factor.grid} dataframe generated by  or}

\item{trunc_label}{Should the output contain a column with printing/reading-friendly, truncuated labels to indicate which factor(s)/levels(s) are being used to filter? Default is TRUE. Mutlple levels of a single factor are separated with a colon: (e.g., <code>fact1level1:fact1level2</code>); factors are separated by a double-colon: (e.g., <code>fact1level1:fact1level2::fact2level1</code>)}

\item{for_aov}{A logical indicating whether the data be returned for passing to a function in @family aov-out functions? Defaults to TRUE.}
}
\value{
a data.frame of factorial-based filtering condition or a  list for passing to additional
}
\description{
Creates a factor grid dataframe of data-filtering conditions for passing to additional to additional factorial-design functions.

A vectorized vesion of the \code{factor_filter} function

Derive filtering conditions from a factor grid
}
\details{
Formula-strings for filtering a data set based on the levels of crossed factors
}
\seealso{
Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{factor_grid}}
}
