% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_grid-function.R
\name{factor_grid}
\alias{factor_grid}
\alias{factor_grid.default}
\alias{factor_grid.list}
\alias{factor_grid.data.frame}
\title{Create a data frame of variable combinations for a fully-crossed factorial design}
\usage{
factor_grid(data, ...)

\method{factor_grid}{default}(data, ...)

\method{factor_grid}{list}(data, ...)

\method{factor_grid}{data.frame}(data, ..., ind_vars)
}
\arguments{
\item{data}{data that will be converted to a factorgrid dataframe}

\item{...}{additional arguments}

\item{ind_vars}{a character vector naming the columns to extract factor levels from}
}
\value{
a dataframe of class "factor.grid" for passing to additional factorial-design functions
}
\description{
This the generic factor_grid function.

A vectorized vesion of the \code{factor_grid} function

Create a factor grid object from a named list of factors

Create a \code{factor.grid} df from a dataframe
}
\seealso{
Other factorial-design functions: \code{\link{combine_factor_levels}},
  \code{\link{factor_filter}}, \code{\link{factor_summary}}
}
