% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_summary-function.R
\name{factor_summary}
\alias{factor_summary}
\title{Summary stats for factors}
\usage{
factor_summary(.data, .dep_var, .ind_var, .grp_var, ...)
}
\arguments{
\item{.data}{for which a summary is to be computed}

\item{.dep_var}{dependent measure to compute summaries over}

\item{.ind_var}{FALSE (default) for collapsing over all variables, or a vector of variables to collapses over in stages}

\item{.grp_var}{a grouping factor}

\item{...}{additional arguments}
}
\value{
a data.frame of summary stats
}
\description{
Summary stats for factors
}
\seealso{
Other factorial-design functions: \code{\link{aov_fgrid}},
  \code{\link{combine_factor_levels}},
  \code{\link{create_fgrid}}, \code{\link{expand_fgrid}},
  \code{\link{summary_form_grid}}
}
