% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_summary-function.R
\name{factor_summary}
\alias{factor_summary}
\title{Summary stats for factors}
\usage{
factor_summary(data, dv, iv, grp_factor)
}
\description{
@importFrom magrittr %>%
 @param data to summarize
 @param dv the dependent variable
 @param iv a vector with the names of the independent variables
 @param grp_factor a grouping factor
 @return a data.frame of summary stats
 @family factorial-design functions
 @rdname factor_summary
 @include summary_stats-function.R
 @include standard_error-function.R
 @include combine_factor_levels-function.R
 @include select_dots-function.R
 @export
}
