% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist_to_dots-function.R
\name{flist_to_dots}
\alias{flist_to_dots}
\title{Create quoted input for the NSE dots argument from list of formulas or a character vector of such a list for the NSE dots arugment.}
\usage{
flist_to_dots(formlist)
}
\arguments{
\item{formlist}{Either a list of comma-separated RHS formulas, as \code{list(~f1, ~f2)}, or a character vector of such a list: \code{"list(~col1,col2)"}.}
}
\value{
A quoted expression that can be interpred as a \code{dots} argument.
}
\description{
\code{flist_to_dots} prepares a list of formulas or a character vector of such a list to be passed to the \code{.dots} argument of the NSE versions of dplyr \code{\link[dplyr]{nse}} functions like \code{\link[dplyr]{select_}}. It can be useful in cases where you've got a list of conditions in a separate dataframe, etc.
}
\examples{
 ## With a list of RHS formulas
filter_list <-
  list(~Sepal.Length < 5.0, ~Petal.Length > 1.3)
filter_list <-
  flist_to_dots(filter_list)

head(
iris \%>\%
    dplyr::filter_(.dots = filter_list)
    )

## With a character vector 'list'
df <-
  dplyr::tibble(
    col1 = 1:2,
    filter_form = c("list(~Sepal.Length,~Petal.Length)", "list(~Sepal.Length)"))

select_form <-
  flist_to_dots(df$filter_form[2])
head(
  iris \%>\%
    dplyr::select_(.dots = select_form)
    )

}

