% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist_to_dots-function.R
\name{flist_to_dots}
\alias{flist_to_dots}
\title{Return a list of formulas or a character vector of such a list for the NSE dots arugment.}
\usage{
flist_to_dots(.formlist)
}
\arguments{
\item{.formlist}{quoteed or unquoted list of comma-separated RHS formulas,}
}
\value{
An expression that can be interpred as a \code{dots} argument.
}
\description{
Return a list of formulas or a character vector of such a list for the NSE dots arugment.
}
\seealso{
sweet_dots
}
