% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range_grt_than-function.R
\name{get_range_grt_than}
\alias{get_range_grt_than}
\title{Return p-value as a range with in indicator for values greater than .10.}
\usage{
get_range_grt_than(p)
}
\arguments{
\item{p}{A number for which a p-value rage should be returned.}
}
\value{
The significance range: e.g., "<.001"
}
\description{
May be useful when non-singificant values are going to be displayed or reported.
}
