% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pround-function.R
\name{pround}
\alias{pround}
\alias{pround.numeric}
\alias{pround.character}
\alias{pround.default}
\title{pround: Helper function for sweetstat}
\usage{
pround(.p, ...)

\method{pround}{numeric}(.p, ..., .rnd_digit = 3)

\method{pround}{character}(.p, ..., .rnd_digit = 3)

\method{pround}{default}(.p, ...)
}
\arguments{
\item{.p}{a number to be rounded}

\item{...}{(ignored) args to be passed to other methods}

\item{.rnd_digit}{number to pass to <code>round</code>; defaults to 3}
}
\value{
a rounded numeric p value
}
\description{
pround: Helper function for sweetstat

Pass a numeric value to pround

Pass a numeric string to pround

Vectorized pround
}
\seealso{
Other sweetstat functions: \code{\link{statround}},
  \code{\link{sweetpround}}, \code{\link{sweetp}},
  \code{\link{sweetround}}, \code{\link{sweetstat}}
}
