% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_dots-function.R
\name{select_dots}
\alias{select_dots}
\alias{select_dots.default}
\alias{select_dots.vector}
\alias{select_dots.list}
\title{A convience function for programming with dplyr's select_}
\usage{
select_dots(data, ...)

\method{select_dots}{default}(data, ...)

\method{select_dots}{vector}(data, ...)

\method{select_dots}{list}(data, ...)
}
\arguments{
\item{data}{data to pass to \code{\link[dplyr]{select_}}'s .dots parameter.}

\item{...}{further arguments passed to or from other methods}
}
\value{
a formula
}
\description{
A convience function for programming with dplyr's select_

Pass a string to select_

Pass a character vector to select_

Pass a list of column names to select_
}
