% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statround-function.R
\name{statround}
\alias{statround}
\alias{statround.numeric}
\alias{statround.character}
\alias{statround.default}
\title{statround: Helper functions for sweetstat}
\usage{
statround(.stat, ...)

\method{statround}{numeric}(.stat, ..., .rnd_digit = 3)

\method{statround}{character}(.stat, ..., .rnd_digit = 3)

\method{statround}{default}(.stat, ...)
}
\arguments{
\item{.stat}{a number to be rounded}

\item{...}{(ignored) args to be passed to other methods}

\item{.rnd_digit}{number to pass to <code>round</code>; defaults to 3}
}
\value{
a rounded numeric
}
\description{
statround: Helper functions for sweetstat

Pass a numeric value to statround

Pass a numeric string to statround

Vectorized statround
}
\seealso{
Other sweetstat functions: \code{\link{pround}},
  \code{\link{sweetpround}}, \code{\link{sweetp}},
  \code{\link{sweetround}}, \code{\link{sweetstat}}
}
