% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats-function.R
\name{summary_stats}
\alias{summary_stats}
\title{Standard summary stats}
\usage{
summary_stats(.data, .cols, .variable)
}
\arguments{
\item{.data}{data for which a summary is to be computed}

\item{.cols}{a character vector of names to be included in summaries}

\item{.variable}{variable over which summary is to be computed}
}
\value{
a data frame with columns for:  N = length, M = mean, SD = sd, SE = standard_error
}
\description{
This is the generic version of the \code{summary_stats}
}
