% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetp-function.R
\name{sweetp}
\alias{sweetp}
\alias{sweetp.default}
\alias{sweetp.numeric}
\alias{sweetp.character}
\title{APA-style decimal rounding for p-vals}
\usage{
sweetp(stat, ...)

\method{sweetp}{default}(stat, ...)

\method{sweetp}{numeric}(stat, not_as_range, snip_lead_zero)

\method{sweetp}{character}(stat, not_as_range, snip_lead_zero)
}
\arguments{
\item{stat}{a statistic to be rounded in APA style}

\item{...}{further arguments passed to or from other methods}

\item{not_as_range}{P values be returned in a range? (APA suggests a range for figures and tables in some cases)}

\item{snip_lead_zero}{Should the leading zero (before the decimal) be remved? APA recommends this for numbers that cannot exceed 1.}
}
\value{
p value stat as APA-rounded character vector
}
\description{
This is the generic version of the function.

A vectorized vesion of the \code{sweetp} function

Sweeten an numeric vector

Sweeten a characer vector
}
\examples{
lapply(runif(10,0.0004,0.034), sweetp)

}
\seealso{
Other sweetstat functions: \code{\link{pround}},
  \code{\link{statround}}, \code{\link{sweetstat}}

Other sweetstat functions: \code{\link{pround}},
  \code{\link{statround}}, \code{\link{sweetstat}}

Other sweetstat functions: \code{\link{pround}},
  \code{\link{statround}}, \code{\link{sweetstat}}

Other sweetstat functions: \code{\link{pround}},
  \code{\link{statround}}, \code{\link{sweetstat}}
}
