% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetpround-function.R
\name{sweetpround}
\alias{sweetpround}
\title{Sweeten a statistic: Helper for Stat Sweet}
\usage{
sweetpround(.stat, rnd_digit, interval = FALSE, lead_zero = FALSE, ...)
}
\arguments{
\item{.stat}{a numeric value representing a statistic}

\item{rnd_digit}{value to pass to round function, defaults to 2}

\item{interval}{(FALSE) Should P values be returned in a range?  APA suggests reporting p values in a range for figures and tables in some cases. Automatically returnd without leading zeros, so if lead_zero is ignored when interval=TRUE}

\item{lead_zero}{(FALSE) Should the leading zero (before the decimal) be preserved? Defaults to FALSE, removing the zero in accordance with APA recommendations.}

\item{...}{(ignored)}
}
\value{
value for passing to sweetstat
}
\description{
Sweeten a statistic: Helper for Stat Sweet
}
\seealso{
Other sweetstat functions: \code{\link{pround}},
  \code{\link{statround}}, \code{\link{sweetp}},
  \code{\link{sweetround}}, \code{\link{sweetstat}}
}
