% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetround-function.R
\name{sweetround}
\alias{sweetround}
\title{Sweeten a statistic: Helper for Stat Sweet}
\usage{
sweetround(.stat, rnd_digit, ...)
}
\arguments{
\item{.stat}{a numeric value representing a statistic}

\item{rnd_digit}{value to pass to round function, defaults to 2}

\item{...}{(ignored)}
}
\value{
value for passing to sweetstat
}
\description{
Sweeten a statistic: Helper for Stat Sweet
}
\seealso{
Other sweetstat functions: \code{\link{pround}},
  \code{\link{statround}}, \code{\link{sweetstat}}
}
