% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetstat-function.R
\name{sweetstat}
\alias{sweetstat}
\alias{sweetstat.default}
\alias{sweetstat.numeric}
\alias{sweetstat.character}
\title{APA-style decimal rounding for stats}
\usage{
sweetstat(stat, ...)

\method{sweetstat}{default}(stat)

\method{sweetstat}{numeric}(stat)

\method{sweetstat}{character}(stat)
}
\arguments{
\item{stat}{a statistic to be rounded in APA style}

\item{...}{further arguments passed to or from other methods}
}
\value{
stat as APA-rounded character vector
}
\description{
This is the generic version of the function.

A vectorized vesion of the \code{sweetstat} function

Sweeten a stat for APA-styling
}
\seealso{
Other sweetstat functions: \code{\link{pround}},
  \code{\link{statround}}, \code{\link{sweetp}}

Other sweetstat functions: \code{\link{pround}},
  \code{\link{statround}}, \code{\link{sweetp}}

Other sweetstat functions: \code{\link{pround}},
  \code{\link{statround}}, \code{\link{sweetp}}

Other sweetstat functions: \code{\link{pround}},
  \code{\link{statround}}, \code{\link{sweetp}}
}
